solving issue with complex method (codefactor)
